fhir_app/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2.1
django-environ==0.10.0
djangorestframework==3.14.0
idna==3.4
jwt==1.3.1
psycopg2==2.9.6
pycparser==2.21
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
